Moved some more stuff around to make it a bit more organized
